Move non-app modules/packages out of app package (#1456)
This PR resolves circular dependencies caused by placing common things inside the app package, e.g., parameters, entry point utilities, etc.
