+++ src/aiidalab_qe/__main__.py
@@ -92,7 +92,7 @@ def download_pseudos(dest):
 @click.option("-p", "--profile", default=_DEFAULT_PROFILE)
 def test_plugin(plugin_name, profile):
     from aiida import load_profile
-    from aiidalab_qe.app.utils import test_plugin_functionality
+    from aiidalab_qe.plugins.utils import test_plugin_functionality
 
     load_profile(profile)
