+++ src/aiidalab_qe/common/process/tree.py
@@ -11,9 +11,9 @@ from aiida.common.exceptions import NotExistent
 from aiida.common.links import LinkType
 from aiida.engine import ProcessState
 from aiida.tools.graph.graph_traversers import traverse_graph
-from aiidalab_qe.app.utils import get_entry_items
 from aiidalab_qe.common.mixins import HasProcess
 from aiidalab_qe.common.mvc import Model
+from aiidalab_qe.plugins.utils import get_entry_items
 from aiidalab_widgets_base import LoadingWidget
 
 from .state import STATE_ICONS
