+++ src/aiidalab_qe/workflows/__init__.py
@@ -8,45 +8,13 @@ from aiida_quantumespresso.common.types import ElectronicType, RelaxType, SpinTy
 from aiida_quantumespresso.data.hubbard_structure import HubbardStructureData
 from aiida_quantumespresso.utils.mapping import prepare_process_inputs
 from aiida_quantumespresso.workflows.pw.relax import PwRelaxWorkChain
+from aiidalab_qe.plugins.utils import get_entry_items
 from aiidalab_qe.utils import enable_pencil_decomposition
 
 XyData = DataFactory("core.array.xy")
 StructureData = DataFactory("core.structure")
 BandsData = DataFactory("core.array.bands")
 Orbital = DataFactory("core.orbital")
-
-
-# functions to get entry points is copied from aiidalab_qe.app.utils
-# because we want to decouple the workflows from the app, so I copied it here
-# instead of importing it.
-# load entry points
-
-
-def get_entries(entry_point_name="aiidalab_qe.property"):
-    from importlib.metadata import entry_points
-
-    entries = {}
-    for entry_point in entry_points().get(entry_point_name, []):
-        try:
-            # Attempt to load the entry point
-            loaded_entry_point = entry_point.load()
-            entries[entry_point.name] = loaded_entry_point
-        except Exception as e:
-            # Handle loading errors
-            print(f"Failed to load entry point {entry_point.name}: {e}")
-
-    return entries
-
-
-# load entry point items
-def get_entry_items(entry_point_name, item_name="workchain"):
-    entries = get_entries(entry_point_name)
-    return {
-        name: entry_point.get(item_name)
-        for name, entry_point in entries.items()
-        if entry_point.get(item_name, False)
-    }
-
 
 plugin_entries = get_entry_items("aiidalab_qe.properties", "workchain")
