+++ tests/test_entry_point.py
@@ -1,5 +1,5 @@
 def test_entry_point():
-    from aiidalab_qe.app.utils import get_entries
+    from aiidalab_qe.plugins.utils import get_entries
 
     entries = get_entries()
     assert "bands" in entries
